test(chat): clean up attachment download tests for streaming refactor
Remove tests for deleted functions (_summarize_thinking, _generate_response)
and keep only the attachment processing tests with updated mocks.
Rename from bot_remaining_gaps_test.py to bot_attachment_download_test.py.
